refactor: minor refactoring
